backend/requirements.txt
asgiref==3.11.0
Django>=5.2,<5.3
djangorestframework==3.16.1
sqlparse==0.5.5
psycopg[binary]>=3.1.0
django-environ>=0.11.0